添加模型请求方式限制

+++ rest_framework/gin/database-sql/dispatcher.go
@@ -21,21 +21,25 @@ func Dispatcher(m ViewAPI, c *gin.Context) {
 		)
 		return
 	}
-	switch c.Request.Method {
-	case "GET":
-		fmt.Println(c.Param("id"))
-		if len(c.Param("id")) == 1 {
-			m.ListViewAPI(c)
-		} else {
-			m.RetrieveViewAPI(c)
+	if inFields(c.Request.Method, m.GetAllowMethod()) {
+		switch c.Request.Method {
+		case "GET":
+			fmt.Println(c.Param("id"))
+			if len(c.Param("id")) == 1 {
+				m.ListViewAPI(c)
+			} else {
+				m.RetrieveViewAPI(c)
+			}
+		case "POST":
+			m.CreateViewAPI(c)
+		case "PUT":
+			m.UpdateViewAPI(c)
+		case "DELETE":
+			m.DeleteViewAPI(c)
+		default:
+			c.JSON(http.StatusBadRequest, gin.H{"code": http.StatusBadRequest, "msg": "请求方式不被允许", "data": ""})
 		}
-	case "POST":
-		m.CreateViewAPI(c)
-	case "PUT":
-		m.UpdateViewAPI(c)
-	case "DELETE":
-		m.DeleteViewAPI(c)
-	default:
+	} else {
 		c.JSON(http.StatusBadRequest, gin.H{"code": http.StatusBadRequest, "msg": "请求方式不被允许", "data": ""})
 	}
 	return
